coding/services/booking 2/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.24.0


# Database
pymysql==1.0.3
mysql-connector-python==8.2.0

# Schema and Validation
pydantic==2.4.2

# GraphQL
strawberry-graphql>=0.171.1

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
requests==2.31.0
python-dotenv==1.0.0
pika==1.3.1 